Remove a number of entries from module denylist
The entries removed from this list are mostly still unconverted because
of missing deps (they are automatically marked unconvertible because of
missing deps). The remainder are actually generated by bp2build as a
result of this change. These cases are for tests that build but don't
pass successfully. We thus simultaneously make a change refining
bp2build.sh. (See topic)

Test: bp2build.sh
Test: m nothing
Test: manual vetting of a number of modules via before/after
analyze_build

Change-Id: Ie9f2c11939258eb3e70ca1e385fe99a3cc4ce93a