Assign default bug component to targets in this directory.
This CL is being created to improve the test attribution in android. It has
been found that the tests owned by this OWNERS are not not being attributed
to a buganiser component.
This is part of a bigger effort to attribute all of the aosp code. 

If you think that the buganiser component is not correct please update
this CL with correct component and drop a +2. We will use the updated component and merge the CL

Bug: 309090038
Change-Id: Ia9dbd8ada1fc9c3ab358c8992d804925f6719f3c
Test: N/A