Add team property to all modules.
This allows vendors (like google) to specify which team owns the test
module and code.

Team is a commonProperty on modules and points to the designate "team"
module.  The DepsMutator adds the dependency on the "team" module and
"GenerateBuildActions" write the team data to intermediate files.

A new singleton rule, all_teams visits all modules and writes out
the proto containing the team for each module.
If a module doesn't have a team, then it finds the package in the
blueprint file and parent directory blueprint files that have a
default_team and uses that team.

*** BRANCH DIFF ***
*** BRANCH DIFF ***
   Some small diffs in writing proto file on udc-branch.
   i.e. write as textproto instead of binary and only write modules
   with teams or default teams from package, not all modules.
   This is because udc doesn't have the newer WriteFile rule from
   raw_files: https://source.corp.google.com/h/android/platform/build/soong/+/df4ce7b565598c5c3421bc6e5942b7894e8ef6fd
*** BRANCH DIFF ***
*** BRANCH DIFF ***

Test: m all_teams
Test: go test ./python ./java ./cc ./rust ./android
Test: added team to HelloWorldHostTest and built the new asciiproto target
Test: added package default_team and checkout output proto.
Change-Id: I5c07bf489de460a04fc540f5fff0394f39f574a7
Merged-In: I5c07bf489de460a04fc540f5fff0394f39f574a7